pass

+++ main.go
@@ -88,7 +88,7 @@ func watch() {
 				}
 				var cpurate int
 				if lastCPUTime, ok := t[name]; ok {
-					cpurate = (float32((info.CpuTime-lastCPUTime)*100) / float32(20*info.NrVirtCpu*10000000))
+					cpurate = int(float32((info.CpuTime-lastCPUTime)*100) / float32(20*info.NrVirtCpu*10000000))
 					if cpurate < 1 {
 						cpurate = 1
 					}
